Return the existing hydrated url if one exists
The problem seems to be that sometimes a call to route() would trigger
multiple calls to .url(), which would attempt to reparse the template
to determine the resulting url string. However, because we're removing
the values from the passed in urlParams, the second attempt to parse
the template would cause a failure - urlParams is empty.

Now, we keep the resulting url string and simply return it allowing
multiple calls to .url() to safely return the same result.